Add methods to support a custom tab widget
This custom tab widget can be a plus button in the tab bar or a set of arrow
keys to scroll the tab bar.

Using a functor as a paremeter allows for maximum flexibility because the user
can decide how the widget is created and how the signals are connected.