Added option to fix resizing of DockWidget and option to hide DockAreaWidget titlebar.
Added option to set DockWidget resize direction. By default the DockWidget can resize in both directions (horizontal and vertical). If the direction is however limited to only one or none and the DockWidget is the only visible dock widget in it's dock area widget, the respective splitter handle will be disabled (if the splitter direction is the same as the non resizing direction of the dock widget). Also added option to hide Titlebar of a DockAreaWidget completely.