Central Widget concept added.
Added option to set a CENTRAl WIDGET into the Dock Manager. It is an important part of a docking system. If a central widget is set, it will be the only widget that will be stretched with the container (window) as the remaining dock widgets will only be resized on vertical direction if docked left or right and in horizontal direction if docked top or bottom. This is a standard behavior of all docking systems where an central widget is inroduced.
Also this update enabled a DockWidget to be user resizable in defined direction only. if a DockWidget is set to be user resizable only horizontaly and is docked at bottom area, the splitter hosting this dock widget  will have the handle bordering this dock widget disabled.
Third update is the option to set a dock widget to have no title bar. If a dock widget is set to have no title bar and is the only widget in a dock area, the area title bar will be hidden.