Don't show empty floating containers on startup
If you add a floating widget to the dock manager, and then close it before the dock manager is shown, showing the dock manager will currently cause the empty floating dock container to be shown.

This fixes this behaviour so that the empty floating container is not shown in this circumstance, making it consistent with the behaviour if the above is performed while the dock manager is visible.