Improve central widget resize behaviour
When using the central widget functionality, only the central widget
is resized when the window is resized. However this behaviour only
happens when there is a single widget in the central area.

This patch adds a new config flag called AllowMultipleCentralWidgets
which allows this resize behaviour to also take effect when there
is more than one widget docked in the central area.